extract GrantType to own file
